route PUT and DELETE ok

+++ controllers/book.js
@@ -1,4 +1,5 @@
 const Book = require('../models/book') //import du model "book"
+const fs = require('fs') //import du module fs de node pour la gestion des fichiers
 
 //ajoute un livre à la BDD
 exports.addBook = async (req, res) => {
@@ -22,8 +23,8 @@ exports.addBook = async (req, res) => {
 			message: 'Livre ajouté à la BDD !',
 		})
 	} catch (error) {
-		console.error('Book (addBook) error : ', error)
 		res.status(400).json({ error })
+		console.error('Book (addBook) error : ', error)
 	}
 }
 
@@ -43,17 +44,58 @@ exports.searchBook = (req, res, next) => {
 
 //modifie un livre à la BDD via son "id"
 exports.updateBook = (req, res) => {
-	Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-		.then(() =>
-			res.status(200).json({ message: 'Le livre a bien été modifié !' })
-		)
+	//si l'on récupère un fichier alors on parse l'objet en modifiant l'url, sinon on récupère le corps de la requête sans le modifier
+	const bookFromFront = req.file
+		? {
+				...JSON.parse(req.body.book), //on parse notre réponse reçue sous format json et on définit la nouvelle url
+				imageUrl: `${req.protocol}://${req.get('host')}/images/${
+					req.file.filename
+				}`,
+		  }
+		: { ...req.body }
+
+	delete bookFromFront.userId //sera supprime l'userId venant de la requête
+
+	Book.findOne({ _id: req.params.id })
+		.then((bookFound) => {
+			//avant de modifier un livre on vérifie si le userId en BDD est différent de celui envoyé par le token dans le corps de la requête
+			if (bookFound.userId !== req.auth.userId) {
+				res.status(401).json({ message: 'Non authorisé' })
+			} else {
+				Book.updateOne(
+					{ _id: req.params.id },
+					{ ...bookFromFront, _id: req.params.id }
+				)
+					.then(() =>
+						res.status(200).json({ message: 'Le livre a bien été modifié !' })
+					)
+					.catch((error) => res.status(400).json({ error }))
+			}
+		})
 		.catch((error) => res.status(400).json({ error }))
 }
 
 //supprime un livre à la BDD via son "id"
 exports.deleteBook = (req, res) => {
-	Book.deleteOne({ _id: req.params.id })
-		.then(() => res.status(200).json({ message: 'Livre supprimé de la BDD !' }))
+	Book.findOne({ _id: req.params.id })
+		.then((bookFound) => {
+			//avant de supprimer un livre on vérifie si le userId en BDD est différent de celui envoyé par le token dans le corps de la requête
+			if (bookFound.userId !== req.auth.userId) {
+				res.status(401).json({ message: 'Non authorisé' })
+			} else {
+				const filename = bookFound.imageUrl.split('/images/')[1] //récupère le chemin de l'image
+
+				//pour la supprimer avec la méthode ".unlick()" de "fs"
+				fs.unlink(`images/${filename}`, () => {
+					//supprime ensuite le livre
+					Book.deleteOne({ _id: req.params.id })
+						.then(() =>
+							res.status(200).json({ message: 'Livre supprimé de la BDD !' })
+						)
+						.catch((error) => res.status(400).json({ error }))
+				})
+			}
+		})
 		.catch((error) => res.status(400).json({ error }))
 }
 
@@ -63,7 +105,7 @@ exports.addNotation = (req, res) => {
 		.then(() =>
 			res
 				.status(200)
-				.json({ message: 'La notation a bien été prise en compte !' })
+				.json({ message: '✅ La notation a bien été prise en compte !' })
 		)
 		.catch((error) => res.status(400).json({ error }))
 }
